Added ability to configure compass angle offset:
- Renamed `zSpeechSettings.h` to `IniSettings.h`
- Added `CompassAngleOffset` option in `gothic.ini`
