+++ YAGA/Compass.cpp
@@ -2,8 +2,6 @@
 // Union SOURCE file
 
 namespace GOTHIC_ENGINE {
-	const int compassAngleOffset = 45;
-
 	float GetCompassAngle() {
 		zVEC3 atVectorWorld3D = player->GetAtVectorWorld();
 		zVEC2 atVectorWorld2D = zVEC2( atVectorWorld3D[VX], atVectorWorld3D[VZ] );
@@ -13,11 +11,11 @@ namespace GOTHIC_ENGINE {
 	void RotateTowardsCompassAngle(int direction) {
 		float currentAngle = GetCompassAngle();
 
-		int remainder = static_cast<int>(currentAngle) % compassAngleOffset;
+		int remainder = static_cast<int>(currentAngle) % Settings::zCompassAngleOffset;
 		if (direction == 1)
-			remainder = compassAngleOffset - remainder;
+			remainder = Settings::zCompassAngleOffset - remainder;
 
-		float offsetAngle = remainder != 0 ? remainder * direction : compassAngleOffset * direction;
+		float offsetAngle = remainder != 0 ? remainder * direction : Settings::zCompassAngleOffset * direction;
 		player->RotateLocalY(offsetAngle);
 	}
 }
