+++ YAGA/Headers.h
@@ -5,11 +5,11 @@
 #pragma region Includes
 #include "UnionHeader.h"
 #include "Utils.h"
+#include "IniSettings.h"
 #include "zSpeechEngine.h"
 #include "zSpeechEngineZhengdu.h"
 #include "zSpeech.h"
 #include "zSpeechEngineNVDA.h"
-#include "zSpeechSettings.h"
 #include "NPCTracker.h"
 #include "AudioHelper.h"
 #include "Compass.h"
